TD-1445: Keep model utilities client-safe
