+++ packages/shared/src/llm-models/llm-model-service.ts
@@ -1,5 +1,6 @@
 import { LlmModelSelectModel } from '../db/schema';
 import { dbGetModelByRoleAndFederalStateId } from '@shared/db/functions/llm-model';
+import { getFirstTextModel } from './llm-model-utils';
 
 export async function getDefaultModel({
   federalStateId,
@@ -21,34 +22,4 @@ export async function getDefaultModelNameByFederalStateId(
   const model = await getDefaultModel({ federalStateId, models });
   if (!model) throw new Error(`No default text model found for federal state ${federalStateId}`);
   return model.name;
-}
-
-/**
- * Get the first text model from a list of models, excluding mistral models.
- * @param models The list of LLM models
- * @returns The first text model or undefined if none found
- */
-export function getFirstTextModel<T extends LlmModelSelectModel>(models: T[]): T | undefined {
-  return models.find(
-    (m) => m.priceMetadata.type === 'text' && !m.name.toLowerCase().includes('mistral'),
-  );
-}
-
-/**
- * Get text models from the list of provided models, optionally excluding mistral models.
- * @param models The list of LLM models
- * @param excludeMistral Whether to exclude mistral models
- * @returns Filtered list of text models
- */
-export function getFilteredTextModels<T extends LlmModelSelectModel>(
-  models: T[],
-  excludeMistral: boolean = false,
-): T[] {
-  let filteredModels = models.filter((m) => m.priceMetadata.type === 'text');
-
-  if (excludeMistral) {
-    filteredModels = filteredModels.filter((m) => !m.name.toLowerCase().includes('mistral'));
-  }
-
-  return filteredModels;
 }
